refactor: Move UrlHistory to /Caching/ folder
