+++ src/Dfe.PlanTech.Application/Caching/Interfaces/IUrlHistory.cs
@@ -1,4 +1,4 @@
-namespace Dfe.PlanTech.Application.Core;
+namespace Dfe.PlanTech.Application.Caching.Interfaces;
 
 /// <summary>
 /// For tracking user's URL history
