+++ src/Dfe.PlanTech.Web/Controllers/QuestionsController.cs
@@ -1,4 +1,4 @@
-using Dfe.PlanTech.Application.Core;
+using Dfe.PlanTech.Application.Caching.Interfaces;
 using Dfe.PlanTech.Application.Questionnaire.Queries;
 using Dfe.PlanTech.Domain.Questionnaire.Models;
 using Dfe.PlanTech.Web.Models;
