+++ src/Dfe.PlanTech.Web/ProgramExtensions.cs
@@ -1,5 +1,5 @@
 using Dfe.PlanTech.Application.Caching.Interfaces;
-using Dfe.PlanTech.Application.Core;
+using Dfe.PlanTech.Application.Caching.Models;
 using Dfe.PlanTech.Domain.Caching.Interfaces;
 using Dfe.PlanTech.Domain.Caching.Models;
 using Dfe.PlanTech.Domain.Content.Models.Options;
