+++ tests/Dfe.PlanTech.Web.UnitTests/Middleware/UrlHistoryMiddlewareTests.cs
@@ -1,4 +1,5 @@
 using Dfe.PlanTech.Application.Caching.Interfaces;
+using Dfe.PlanTech.Application.Caching.Models;
 using Dfe.PlanTech.Application.Core;
 using Dfe.PlanTech.Domain.Caching.Models;
 using Dfe.PlanTech.Web.Helpers;
@@ -32,7 +33,7 @@ public class UrlHistoryMiddlewareTests
     }
 
     [Fact]
-    public void Should_PopHistory_When_Navigating_Backwards()
+    public async Task Should_PopHistory_When_Navigating_Backwards()
     {
         var requestMock = new Mock<HttpRequest>();
         requestMock.SetupGet(request => request.Host).Returns(new HostString("www.website.com"));
@@ -45,7 +46,7 @@ public class UrlHistoryMiddlewareTests
         var urlHistory = new UrlHistoryMiddleware(nextDelegateMock.Object);
 
         var history = new UrlHistory(_cacher);
-        urlHistory.InvokeAsync(contextMock.Object, history);
+        await urlHistory.InvokeAsync(contextMock.Object, history);
 
         var historyCache = history.History;
         Assert.Equal(2, historyCache.Count);
@@ -53,7 +54,7 @@ public class UrlHistoryMiddlewareTests
     }
 
     [Fact]
-    public void Should_AddToHistory_When_Navigating_ToNewPage()
+    public async Task Should_AddToHistory_When_Navigating_ToNewPage()
     {
         var requestMock = new Mock<HttpRequest>();
         requestMock.SetupGet(request => request.IsHttps).Returns(false);
@@ -70,7 +71,7 @@ public class UrlHistoryMiddlewareTests
         var urlHistory = new UrlHistoryMiddleware(nextDelegateMock.Object);
 
         var history = new UrlHistory(_cacher);
-        urlHistory.InvokeAsync(contextMock.Object, history);
+        await urlHistory.InvokeAsync(contextMock.Object, history);
 
         var historyCache = history.History;
